+++ src/main/java/com/atguigu/service/impl/CoachScheduleServiceImpl.java
@@ -16,7 +16,7 @@ public class CoachScheduleServiceImpl implements CoachScheduleService {
 
     @Override
     public boolean updatelesson(String mem_id, String real_club, String real_coach,
-                                String sale_id,String kc_id, Byte seq_no, Date start_time_1,
+                                String sale_id,String kc_id, Integer seq_no, Date start_time_1,
                                 Date end_time_1,String bz1) {
 
         MemberLesson memberLesson = new MemberLesson();
